move LotesList component into dir. add LotesListItem and LotesListItemObra as components
