+++ src/native/components/Lotes/LotesList.js
@@ -1,0 +1,65 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { StyleSheet, SectionList, ListItem, FlatList, TouchableOpacity, RefreshControl, Image, View } from 'react-native';
+import { Card, CardItem, Body, Text, Button } from 'native-base';
+import { Actions } from 'react-native-router-flux';
+
+import Loading from '../Loading';
+import Error from '../Error';
+import Spacer from '../Spacer';
+
+import LotesListItem from './LotesListItem';
+
+const LotesList = ({
+  error,
+  loading,
+  lotes,
+  reFetch,
+}) => {
+  // Loading
+  if (loading) return <Loading />;
+
+  // Error
+  if (error) return <Error content={error} />;
+
+  const keyExtractor = item => item.id;
+
+  const onPress = item => Actions.lote({ match: { params: { id: String(item.id) } } });
+
+  const styles = StyleSheet.create({
+    bordered: {
+      borderBottomWidth: 1,
+      borderStyle: 'solid',
+      borderBottomColor: 'black',
+    }
+  });
+
+  return (
+    <FlatList
+      numColumns={1}
+      data={lotes}
+      renderItem={LotesListItem}
+      keyExtractor={keyExtractor}
+      refreshControl={
+        <RefreshControl
+          refreshing={loading}
+          onRefresh={reFetch}
+        />
+      }
+    />
+  );
+};
+
+LotesList.propTypes = {
+  error: PropTypes.string,
+  loading: PropTypes.bool.isRequired,
+  lotes: PropTypes.array.isRequired,
+  reFetch: PropTypes.func,
+};
+
+LotesList.defaultProps = {
+  error: null,
+  reFetch: null,
+};
+
+export default LotesList;
